Revert "Use different accumulator for earn (#1395)"
This reverts commit cf009647e6f6f57ba7469bb0a4a851eac00f9b10.
