+++ x/incentive/keeper/adapters/swap/adapter_test.go
@@ -92,7 +92,7 @@ func (suite *SwapAdapterTestSuite) TestSwapAdapter_OwnerSharesBySource() {
 		sdk.ZeroDec(),
 	))
 
-	err := suite.app.FundAccount(
+	suite.app.FundAccount(
 		suite.ctx,
 		suite.addrs[0],
 		sdk.NewCoins(
@@ -100,9 +100,7 @@ func (suite *SwapAdapterTestSuite) TestSwapAdapter_OwnerSharesBySource() {
 			sdk.NewCoin(poolDenomB, sdk.NewInt(1000000000000)),
 		),
 	)
-	suite.NoError(err)
-
-	err = suite.app.FundAccount(
+	suite.app.FundAccount(
 		suite.ctx,
 		suite.addrs[1],
 		sdk.NewCoins(
@@ -110,9 +108,8 @@ func (suite *SwapAdapterTestSuite) TestSwapAdapter_OwnerSharesBySource() {
 			sdk.NewCoin(poolDenomB, sdk.NewInt(1000000000000)),
 		),
 	)
-	suite.NoError(err)
 
-	err = swapKeeper.Deposit(
+	err := swapKeeper.Deposit(
 		suite.ctx,
 		suite.addrs[0],
 		sdk.NewCoin(poolDenomA, sdk.NewInt(100)),
@@ -222,22 +219,22 @@ func (suite *SwapAdapterTestSuite) TestSwapAdapter_TotalSharesBySource() {
 		sdk.ZeroDec(),
 	))
 
-	suite.NoError(suite.app.FundAccount(
+	suite.app.FundAccount(
 		suite.ctx,
 		suite.addrs[0],
 		sdk.NewCoins(
 			sdk.NewCoin(poolDenomA, sdk.NewInt(1000000000000)),
 			sdk.NewCoin(poolDenomB, sdk.NewInt(1000000000000)),
 		),
-	))
-	suite.NoError(suite.app.FundAccount(
+	)
+	suite.app.FundAccount(
 		suite.ctx,
 		suite.addrs[1],
 		sdk.NewCoins(
 			sdk.NewCoin(poolDenomA, sdk.NewInt(1000000000000)),
 			sdk.NewCoin(poolDenomB, sdk.NewInt(1000000000000)),
 		),
-	))
+	)
 
 	err := swapKeeper.Deposit(
 		suite.ctx,
