+++ x/incentive/types/params.go
@@ -133,10 +133,6 @@ func (p Params) Validate() error {
 		return err
 	}
 
-	if err := validatedRewardPeriodsParam(p.RewardPeriods); err != nil {
-		return err
-	}
-
 	return nil
 }
